+++ src/ts/inputs/inputMaps.ts
@@ -1,7 +1,7 @@
 import { InputMap } from "./inputMap";
 import { SpaceShipControlsInputs } from "../spaceship/spaceShipControlsInputs";
-import { CharacterInputs } from "../spacelegs/characterControlsInputs";
-import { DefaultControlsInputs } from "../defaultController/defaultControlsInputs";
+import { CharacterInputs } from "../characterControls/characterControlsInputs";
+import { DefaultControlsInputs } from "../defaultControls/defaultControlsInputs";
 import { GeneralInputs } from "./generalInputs";
 import { StarSystemInputs } from "./starSystemInputs";
 import { StarMapInputs } from "./starMapInputs";
