+++ src/ts/stellarObjects/star/starModel.ts
@@ -22,7 +22,6 @@ import { getRgbFromTemperature } from "../../utils/specrend";
 import { Settings } from "../../settings";
 import { getOrbitalPeriod } from "../../orbit/orbit";
 import { OrbitProperties } from "../../orbit/orbitProperties";
-import { BodyType, GenerationSteps } from "../../model/common";
 import { StellarType } from "../common";
 import { StarPhysicalProperties } from "../../architecture/physicalProperties";
 import { CelestialBodyModel } from "../../architecture/celestialBody";
@@ -30,6 +29,8 @@ import { wheelOfFortune } from "../../utils/wheelOfFortune";
 import { StellarObjectModel } from "../../architecture/stellarObject";
 import { Color3 } from "@babylonjs/core/Maths/math.color";
 import { RingsModel } from "../../rings/ringsModel";
+import { BodyType } from "../../architecture/bodyType";
+import { GenerationSteps } from "../../utils/generationSteps";
 
 export class StarModel implements StellarObjectModel {
     readonly bodyType = BodyType.STAR;
